Refactor webrpc not panic but return error
